wb-bot/requirements.txt
aiogram==3.22.0
httpx==0.28.1
python-dotenv==1.1.1
openpyxl==3.1.5
